fix tap migration renames with API
We don't load casks and formulae that used to exist in an external
tap but have been moved and renamed to an API tap when using the
API. Loading works correctly when the core formula or cask tap
is tapped locally though.

This PR adds some logic to map tap migration renames to the API
and check that when loading formulae and casks.